Add files via upload
This contract allows to set a string (aka a name) of your choice through the "set" function. "SetSomething" is what I decided to show in the "set" empty box. 
Once something is set, press "value" to return and visualise it. Always remember to compile first, then deploy. Then, functions will be available.