Add files via upload
Smart contract that changes a state from low to high, and has also a function to check if the high state is true or false.
enum is a numerical state. In this case, low corresponds to 0 and High to 1. When Activate is pressed, the state changes from 0 (low) to 1 (high).
Bool expresses true or false states.